+++ src/DoctrineExtensions.php
@@ -28,7 +28,7 @@ final class DoctrineExtensions
     /**
      * Current version of extensions
      */
-    public const VERSION = '3.12.0';
+    public const VERSION = '3.13.0';
 
     /**
      * Hooks all extension metadata mapping drivers into